requirements.txt
python-dateutil~=2.8.2
python-dotenv~=0.19.0
telegram~=0.0.1
pysqlcipher3~=1.0.4